+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 * None.
 
 ### Enhancements
-* None.
+* Add support for in-memory Realms.
+* [Sync] Added support for API key authentication. (Issue [#432](https://github.com/realm/realm-kotlin/issues/432))
 
 ### Fixed
 * None.
@@ -22,6 +23,7 @@
 
 ### Internal
 * Updated to use hierarchical multi platform project structure.
+* Updated to Realm Core 12.10.0, commit 8ce82fe3a8d5a2fbc89d719de8559f5a792c2dc9.
 
 
 ## 1.4.0 (2022-10-17)